BROKEN/MINOR: listeners: move fd_stop_recv() to the receiver's socket code
BROKEN: we first need to undo f2cb169487 : fd_update_cache() doesn't exist
anymore and we have updt_fd_list() instead which allows a thread to update
another one's FD state. This is mandatory for resume which needs to move
the fd_want_recv() at the same place as the call to listen(), which should
be in proto->resume(l).

fd_stop_recv() has nothing to do in the generic listener code, it's per
protocol as some don't need it. For instance with abns@ it could even
lead to fd_stop_recv(-1). And later with QUIC we don't want to touch
the fd at all!